+++ api/CHANGELOG.md
@@ -13,6 +13,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Derive `Partial, Ord` for `Composition`.
 - Derive `Partial, Ord, Hash` for `Term`.
 - Derive `Hash` for `Unit`.
+- New feature flag: `wrappable`. Adds the `AsWrappedRef` trait to make it easier for downstream
+  crates to create new types based on `Measurement`, `Unit`, etc. from this crate.
 
 ### Changed